backend/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.29
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
requests==2.28.2
Flask-Cors==5.0.1
Werkzeug==2.2.3
flask-login==0.6.2
